# packages/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SMB Booking API
  version: 1.0.0
servers:
  - url: https://api.example.com
paths:
  /v1/services:
    get:
      summary: List services
      parameters:
        - in: query
          name: accountId
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
  /v1/availability:
    get:
      summary: Get availability
      description: >
        Returns `{ slots: [] }` unless accountId, provider, calendarId,
        start, and end are all supplied, in which case it queries the
        connected calendar (via the Python connectors) for real busy
        intervals and computes free slots. accountId must have a
        provider calendar already connected via
        `/oauth/{provider}/start` + `/oauth/{provider}/callback`.
      parameters:
        - in: query
          name: accountId
          schema: { type: string }
        - in: query
          name: serviceId
          schema: { type: string }
        - in: query
          name: provider
          schema: { type: string, enum: [google, microsoft] }
        - in: query
          name: calendarId
          schema: { type: string }
        - in: query
          name: start
          schema: { type: string, format: date-time }
        - in: query
          name: end
          schema: { type: string, format: date-time }
        - in: query
          name: slotMinutes
          schema: { type: integer, default: 30 }
        - in: query
          name: tz
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      type: object
                      properties:
                        start: { type: string, format: date-time }
                        end: { type: string, format: date-time }
        '400':
          description: Unsupported provider, or accountId has no calendar connected for provider
        '502':
          description: Failed to reach the calendar provider
  /v1/appointments:
    post:
      summary: Create appointment
      security:
        - hmac: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '200':
          description: >
            Replayed an existing appointment for this Idempotency-Key
            (no new calendar event or DB row was created).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '400':
          description: Validation error, unsupported provider, or accountId has no calendar connected for provider
        '409':
          description: Requested time conflicts with a busy interval on the connected calendar
        '502':
          description: Failed to reach the calendar provider
  /oauth/{provider}/start:
    get:
      summary: Begin the OAuth authorization-code flow for a calendar provider
      description: 302-redirects the browser to the provider's consent screen.
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string, enum: [google, microsoft] }
        - in: query
          name: accountId
          required: true
          schema: { type: string }
      responses:
        '302':
          description: Redirect to the provider's authorization URL
        '400':
          description: Missing accountId
        '404':
          description: Unsupported provider
  /oauth/{provider}/callback:
    get:
      summary: OAuth callback -- exchanges the authorization code for tokens
      description: >
        The provider redirects here with `code` and `state` after the
        user consents. Tokens are exchanged and stored encrypted
        (AES-256-GCM, key from TOKEN_ENCRYPTION_KEY) in Postgres, keyed
        by (accountId, provider).
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string, enum: [google, microsoft] }
        - in: query
          name: code
          schema: { type: string }
        - in: query
          name: state
          schema: { type: string }
      responses:
        '200':
          description: Calendar connected
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountId: { type: string }
                  provider: { type: string }
                  connected: { type: boolean }
                  expiresAt: { type: string, format: date-time }
        '400':
          description: Missing/invalid code or state
        '404':
          description: Unsupported provider
        '502':
          description: Token exchange with the provider failed
  /v1/appointments/{id}/cancel:
    post:
      summary: Cancel appointment
      security:
        - hmac: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
  /v1/appointments/{id}/reschedule:
    post:
      summary: Reschedule appointment
      security:
        - hmac: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newStartTime: { type: string, format: date-time }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
components:
  securitySchemes:
    hmac:
      type: apiKey
      in: header
      name: X-Signature
  schemas:
    Service:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        duration_min: { type: integer }
        buffer_before_min: { type: integer }
        buffer_after_min: { type: integer }
    AppointmentRequest:
      type: object
      required: [accountId, serviceId, startTime, customer]
      properties:
        accountId: { type: string }
        serviceId: { type: string }
        startTime: { type: string, format: date-time }
        customer:
          type: object
          required: [name, email]
          properties:
            name: { type: string }
            email: { type: string, format: email }
            phone: { type: string }
        notes: { type: string }
        source: { type: string, enum: [human, agent, api] }
        metadata: { type: object, additionalProperties: true }
        provider:
          type: string
          enum: [google, microsoft]
          description: >
            When set, the appointment is checked against and written to
            the connected calendar (accountId must already be connected
            via /oauth/{provider}/start). calendarId becomes required.
        calendarId: { type: string }
        durationMinutes: { type: integer, default: 30 }
    AppointmentResponse:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [requested, tentative, confirmed, canceled, no_show] }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        provider_event_id: { type: string }
        links:
          type: object
          properties:
            ical: { type: string, format: uri }
            manage: { type: string, format: uri }